move write amplify meta into the multi-env-manager
